+++ cypress/integration/main/Encounters.spec.ts
@@ -22,8 +22,8 @@ describe('Encounters', () => {
     cy.get('[data-testid=filter] > input')
       .type('Slumbering Weald')
       .should('have.value', 'Slumbering Weald');
-    cy.get('[data-testid="remove-tooltip"]').click();
     cy.get('[data-testid=pokemon-1]').click();
+    cy.get('[data-testid="remove-tooltip"]').click();
     cy.get('[data-testid=poke-Butterfree]').click();
     cy.get('[data-testid=filter-button]').click();
     cy.get('[data-testid=filter-gen-1]').click();